requirements.txt
groq==0.9.0
openai>=1.0.0
python-dotenv